ensf614-project/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
Faker==19.6.1
pandas==2.0.1
